feat: include shelf power usage in "Average Power/Used TB" and "Average IOPs/Watt" panels
Fixes: #3703

Thanks to @BrendonA667 for reporting!
